parser/requirements.txt
# ─────────────────────────────────────────────────────────────────
#  Parser API — Python Dependencies
#  Pinned for reproducibility. Update with: pip-compile requirements.in
# ─────────────────────────────────────────────────────────────────

# ── Web framework ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9       # required for file upload endpoints
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database ──────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9        # binary avoids libpq build dep in image
alembic==1.13.1               # database migrations

# ── CV text extraction ────────────────────────────────────────────
pdfplumber==0.11.0            # text-layer PDF extraction
pytesseract==0.3.10           # OCR fallback for scanned/image CVs
Pillow==10.3.0                # image processing for pytesseract

# ── NLP / HuggingFace ────────────────────────────────────────────
transformers==4.41.1
torch==2.3.0                  # CPU-only — no CUDA dependency
tokenizers==0.19.1
huggingface-hub==0.23.2
sentencepiece==0.2.0          # required by some tokenisers

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0                 # async HTTP client for health checks
structlog==24.1.0             # structured JSON logging
